modularize cast, setdiff1dAsync, print, buffer
