+++ tfjs-core/src/gradients/Acos_grad.ts
@@ -17,7 +17,7 @@
 
 import {Acos} from '../kernel_names';
 import {GradConfig} from '../kernel_registry';
-import {cast} from '../ops/array_ops';
+import {cast} from '../ops/cast';
 import {div} from '../ops/div';
 import {neg} from '../ops/neg';
 import {square} from '../ops/square';
