+++ tfjs-core/src/gradients/Atan_grad.ts
@@ -18,7 +18,7 @@
 import {Atan} from '../kernel_names';
 import {GradConfig} from '../kernel_registry';
 import {add} from '../ops/add';
-import {cast} from '../ops/array_ops';
+import {cast} from '../ops/cast';
 import {div} from '../ops/div';
 import {square} from '../ops/square';
 import {Tensor} from '../tensor';
