+++ tfjs-core/src/gradients/Cos_grad.ts
@@ -17,7 +17,7 @@
 
 import {Cos} from '../kernel_names';
 import {GradConfig} from '../kernel_registry';
-import {cast} from '../ops/array_ops';
+import {cast} from '../ops/cast';
 import {mul} from '../ops/mul';
 import {neg} from '../ops/neg';
 import {sin} from '../ops/sin';
