+++ tfjs-core/src/gradients/Minimum_grad.ts
@@ -17,7 +17,7 @@
 
 import {Minimum} from '../kernel_names';
 import {GradConfig} from '../kernel_registry';
-import {cast} from '../ops/array_ops';
+import {cast} from '../ops/cast';
 import {greater} from '../ops/greater';
 import {lessEqual} from '../ops/less_equal';
 import {mul} from '../ops/mul';
