+++ tfjs-core/src/gradients/Multiply_grad.ts
@@ -17,8 +17,8 @@
 
 import {Multiply} from '../kernel_names';
 import {GradConfig} from '../kernel_registry';
-import {cast} from '../ops/array_ops';
 import {assertAndGetBroadcastShape, getReductionAxes} from '../ops/broadcast_util';
+import {cast} from '../ops/cast';
 import {mul} from '../ops/mul';
 import {reshape} from '../ops/reshape';
 import {sum} from '../ops/sum';
