+++ tfjs-core/src/gradients/Relu_grad.ts
@@ -16,7 +16,7 @@
  */
 import {Relu} from '../kernel_names';
 import {GradConfig} from '../kernel_registry';
-import {cast} from '../ops/array_ops';
+import {cast} from '../ops/cast';
 import {mul} from '../ops/mul';
 import {step} from '../ops/unary_ops';
 import {Tensor} from '../tensor';
