+++ tfjs-core/src/gradients/Sinh_grad.ts
@@ -17,7 +17,7 @@
 
 import {Sinh} from '../kernel_names';
 import {GradConfig} from '../kernel_registry';
-import {cast} from '../ops/array_ops';
+import {cast} from '../ops/cast';
 import {cosh} from '../ops/cosh';
 import {mul} from '../ops/mul';
 import {Tensor} from '../tensor';
