+++ tfjs-core/src/ops/avg_pool.ts
@@ -24,7 +24,7 @@ import {convertToTensor} from '../tensor_util_env';
 import {TensorLike} from '../types';
 import * as util from '../util';
 
-import {cast} from './array_ops';
+import {cast} from './cast';
 import * as conv_util from './conv_util';
 import {op} from './operation';
 import {reshape} from './reshape';
