+++ tfjs-core/src/ops/minimum.ts
@@ -23,9 +23,10 @@ import {makeTypesMatch} from '../tensor_util';
 import {convertToTensor} from '../tensor_util_env';
 import {TensorLike} from '../types';
 
-import {cast} from './array_ops';
 import {assertAndGetBroadcastShape} from './broadcast_util';
+import {cast} from './cast';
 import {op} from './operation';
+
 
 /**
  * Returns the min of a and b (`a < b ? a : b`) element-wise.
