+++ tfjs-core/src/ops/moments.ts
@@ -20,13 +20,14 @@ import {convertToTensor} from '../tensor_util_env';
 import {TensorLike} from '../types';
 import {parseAxisParam} from '../util';
 
-import {cast} from './array_ops';
 import {expandShapeToKeepDim} from './axis_util';
+import {cast} from './cast';
 import {mean} from './mean';
 import {op} from './operation';
 import {reshape} from './reshape';
 import {square} from './square';
 import {sub} from './sub';
+
 
 /**
  * Calculates the mean and variance of `x`. The mean and variance are
