+++ tfjs-core/src/ops/ops.ts
@@ -39,6 +39,8 @@ export {batchNorm2d} from './batchnorm2d';
 export {batchNorm3d} from './batchnorm3d';
 export {batchNorm4d} from './batchnorm4d';
 export {broadcastTo} from './broadcast_to';
+export {buffer} from './buffer';
+export {cast} from './cast';
 export {ceil} from './ceil';
 export {clone} from './clone';
 export {complex} from './complex';
@@ -108,6 +110,7 @@ export {pad4d} from './pad4d';
 export {pool} from './pool';
 export {pow} from './pow';
 export {prelu} from './prelu';
+export {print} from './print';
 export {prod} from './prod';
 export {rand} from './rand';
 export {randomGamma} from './random_gamma';
@@ -124,6 +127,7 @@ export {reverse3d} from './reverse_3d';
 export {reverse4d} from './reverse_4d';
 export {selu} from './selu';
 export {separableConv2d} from './separable_conv2d';
+export {setdiff1dAsync} from './setdiff1d_async';
 export {sign} from './sign';
 export {sin} from './sin';
 export {sinh} from './sinh';
@@ -153,7 +157,6 @@ export * from './boolean_mask';
 export * from './unary_ops';
 export * from './compare';
 export * from './binary_ops';
-export * from './array_ops';
 export * from './tensor_ops';
 export * from './transpose';
 export * from './softmax';
