+++ tfjs-core/src/ops/setdiff1d_async.ts
@@ -1,6 +1,6 @@
 /**
  * @license
- * Copyright 2018 Google LLC. All Rights Reserved.
+ * Copyright 2020 Google Inc. All Rights Reserved.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -14,45 +14,10 @@
  * limitations under the License.
  * =============================================================================
  */
-
-import {ENGINE} from '../engine';
 import {Tensor, TensorBuffer} from '../tensor';
 import {convertToTensor} from '../tensor_util_env';
-import {DataType, DataTypeMap, Rank, ShapeMap, TensorLike} from '../types';
+import {TensorLike} from '../types';
 import * as util from '../util';
-
-import {op} from './operation';
-
-/**
- * Casts a `tf.Tensor` to a new dtype.
- *
- * ```js
- * const x = tf.tensor1d([1.5, 2.5, 3]);
- * tf.cast(x, 'int32').print();
- * ```
- * @param x The input tensor to be casted.
- * @param dtype The dtype to cast the input tensor to.
- */
-/** @doc {heading: 'Tensors', subheading: 'Transformations'} */
-function cast_<T extends Tensor>(x: T|TensorLike, dtype: DataType): T {
-  const $x = convertToTensor(x, 'x', 'cast');
-
-  // Sanity checks.
-  if (!util.isValidDtype(dtype)) {
-    throw new Error(`Failed to cast to unknown dtype ${dtype}`);
-  }
-  if (dtype === 'string' && $x.dtype !== 'string' ||
-      dtype !== 'string' && $x.dtype === 'string') {
-    throw new Error('Only strings can be casted to strings');
-  }
-
-  const grad = (dy: T) => {
-    return {x: () => dy.clone()};
-  };
-  const attrs = {dtype};
-  return ENGINE.runKernelFunc(
-      backend => backend.cast($x, dtype), {x: $x}, grad, 'Cast', attrs);
-}
 
 /**
  * Computes the difference between two lists of numbers.
@@ -120,59 +85,4 @@ async function setdiff1dAsync_(
   }
   return [buffer.toTensor(), indices.toTensor()];
 }
-
-/**
- * Creates an empty `tf.TensorBuffer` with the specified `shape` and `dtype`.
- *
- * The values are stored in CPU as `TypedArray`. Fill the buffer using
- * `buffer.set()`, or by modifying directly `buffer.values`.
- *
- * When done, call `buffer.toTensor()` to get an immutable `tf.Tensor` with
- * those values.
- *
- * ```js
- * // Create a buffer and set values at particular indices.
- * const buffer = tf.buffer([2, 2]);
- * buffer.set(3, 0, 0);
- * buffer.set(5, 1, 0);
- *
- * // Convert the buffer back to a tensor.
- * buffer.toTensor().print();
- * ```
- *
- * @param shape An array of integers defining the output tensor shape.
- * @param dtype The dtype of the buffer. Defaults to 'float32'.
- * @param values The values of the buffer as `TypedArray`. Defaults to
- * zeros.
- */
-/** @doc {heading: 'Tensors', subheading: 'Creation'} */
-export function buffer<R extends Rank, D extends DataType = 'float32'>(
-    shape: ShapeMap[R], dtype: D = 'float32' as D,
-    values?: DataTypeMap[D]): TensorBuffer<R, D> {
-  dtype = dtype || 'float32' as D;
-  util.assertNonNegativeIntegerDimensions(shape);
-  return new TensorBuffer<R, D>(shape, dtype, values);
-}
-
-/**
- * Prints information about the `tf.Tensor` including its data.
- *
- * ```js
- * const verbose = true;
- * tf.tensor2d([1, 2, 3, 4], [2, 2]).print(verbose);
- * ```
- * @param x The tensor to be printed.
- * @param verbose Whether to print verbose information about the ` Tensor`,
- * including dtype and size.
- */
-/** @doc {heading: 'Tensors', subheading: 'Creation'} */
-function print<T extends Tensor>(x: T, verbose = false): void {
-  console.log(x.toString(verbose));
-}
-
-export {
-  print  // Not wrapped in op() since no need to increase stack trace.
-};
-
-export const cast = op({cast_});
 export const setdiff1dAsync = setdiff1dAsync_;
